safe-cook/requirements.txt
opencv-contrib-python
imagezmq
imutils
websocket-client
flask-socketio==4.3.1
python-socketio==4.6.0
python-engineio==3.13.2
numpy